test(textarea): migrate events
